refactoring. Move ExtractPubXY, AuthClaimFromPubKey
