+++ testvectorgen/utils/utilsv2.go
@@ -327,6 +327,35 @@ func getNodeAuxValue(p *merkletree.Proof) NodeAuxValue {
 	}
 }
 
+func ExtractPubXY(privKHex string) (key *babyjub.PrivateKey, x, y *big.Int) {
+	// Extract pubKey
+	var k babyjub.PrivateKey
+	if _, err := hex.Decode(k[:], []byte(privKHex)); err != nil {
+		panic(err)
+	}
+	pk := k.Public()
+	return &k, pk.X, pk.Y
+}
+
+func AuthClaimFromPubKey(X, Y *big.Int) (*core.Claim, error) {
+	var schemaHash core.SchemaHash
+	//schemaEncodedBytes, _ := hex.DecodeString("ca938857241db9451ea329256b9c06e5") // V1
+	schemaEncodedBytes, _ := hex.DecodeString("013fd3f623559d850fb5b02ff012d0e2") // V2
+	copy(schemaHash[:], schemaEncodedBytes)
+
+	// NOTE: We take nonce as hash of public key to make it random
+	// We don't use random number here because this test vectors will be used for tests
+	// and have randomization inside tests is usually a bad idea
+	revNonce, err := poseidon.Hash([]*big.Int{X})
+	if err != nil {
+		return nil, err
+	}
+
+	return core.NewClaim(schemaHash,
+		core.WithIndexDataInts(X, Y),
+		core.WithRevocationNonce(revNonce.Uint64()))
+}
+
 func SaveTestVector(t *testing.T, fileName string, data string) {
 	t.Helper()
 	path := "testdata/" + fileName + ".json"
